+++ base/remote/NEWS.md
@@ -1,5 +1,7 @@
-# PEcAn.remote 1.7.2.9000
+# PEcAn.remote 1.8.0.9000
 
+
+# PEcAn.remote 1.8.0
 
 ## Added